Fix an issue with the worker's security group
It seems that even when providing a custom security group, Hypershift
always creates a default security group for the NodePool. With the
newest release of Hypershift/OCP, the cloud-controller refuses to sync the
default router's LoadBalancer if it finds 2 security groups with the
kubernetes.io/cluster/cluster-name=owned tag. This causes a failure
during cluster provision as the default router is never reachable.